Move proxier cleanup code into separate files.
(No code changes, just moving between files.)
